+++ spec/starcall/static_data_spec.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../../lib/starcall/static_data'
+require_relative '../../lib/starcall/error'
 
 RSpec.describe Starcall::StaticData do
   before do
@@ -19,7 +20,7 @@ RSpec.describe Starcall::StaticData do
   end
 
   describe '.season_ids' do
-    subject(:season_ids) { described_class.season_ids }
+    subject(:season_ids) { described_class.new.season_ids }
     let(:url) do
       { url: 'http://static.developer.riotgames.com/docs/lol/seasons.json' }
     end
@@ -31,7 +32,7 @@ RSpec.describe Starcall::StaticData do
   end
 
   describe '.queue_ids' do
-    subject(:queue_ids) { described_class.queue_ids }
+    subject(:queue_ids) { described_class.new.queue_ids }
     let(:url) do
       { url: 'http://static.developer.riotgames.com/docs/lol/queues.json' }
     end
@@ -43,7 +44,7 @@ RSpec.describe Starcall::StaticData do
   end
 
   describe '.map_ids' do
-    subject(:map_ids) { described_class.map_ids }
+    subject(:map_ids) { described_class.new.map_ids }
     let(:url) do
       { url: 'http://static.developer.riotgames.com/docs/lol/maps.json' }
     end
@@ -55,7 +56,7 @@ RSpec.describe Starcall::StaticData do
   end
 
   describe '.game_modes' do
-    subject(:game_modes) { described_class.game_modes }
+    subject(:game_modes) { described_class.new.game_modes }
     let(:url) do
       { url: 'http://static.developer.riotgames.com/docs/lol/gameModes.json' }
     end
@@ -67,7 +68,7 @@ RSpec.describe Starcall::StaticData do
   end
 
   describe '.game_types' do
-    subject(:game_types) { described_class.game_types }
+    subject(:game_types) { described_class.new.game_types }
     let(:url) do
       { url: 'http://static.developer.riotgames.com/docs/lol/gameTypes.json' }
     end
@@ -79,7 +80,7 @@ RSpec.describe Starcall::StaticData do
   end
 
   describe '.dd_versions' do
-    subject(:dd_versions) { described_class.dd_versions }
+    subject(:dd_versions) { described_class.new.dd_versions }
     let(:url) do
       { url: 'https://ddragon.leagueoflegends.com/api/versions.json' }
     end
@@ -90,137 +91,15 @@ RSpec.describe Starcall::StaticData do
     end
   end
 
-  context 'region is valid' do
-    before do
-      allow(Starcall::Regions).to receive(:valid?).and_return(true)
-    end
-
-    describe '.dd_region_versions' do
-      subject(:dd_region_versions) { described_class.dd_region_versions }
-      let(:url) do
-        { url: 'https://ddragon.leagueoflegends.com/realms/euw.json' }
-      end
-
-      it 'calls the api' do
-        dd_region_versions
-        expect(ApiRequests).to have_received(:make_request).with(url)
-      end
-    end
-
-    describe '.dd_champions' do
-      subject(:dd_champions) { described_class.dd_champions }
-      let(:url) do
-        { url: "http://ddragon.leagueoflegends.com/cdn/#{response['n']['champion']}/data/en_GB/champion.json" }
-      end
-
-      it 'calls the api' do
-        dd_champions
-        expect(ApiRequests).to have_received(:make_request).with(url)
-      end
-    end
-
-    describe '.dd_items' do
-      subject(:dd_items) { described_class.dd_items }
-      let(:url) do
-        { url: "http://ddragon.leagueoflegends.com/cdn/#{response['n']['item']}/data/en_GB/item.json" }
-      end
-
-      it 'calls the api' do
-        dd_items
-        expect(ApiRequests).to have_received(:make_request).with(url)
-      end
-    end
-
-    describe '.dd_summoner_spells' do
-      subject(:dd_summoner_spells) { described_class.dd_summoner_spells }
-      let(:url) do
-        { url: "http://ddragon.leagueoflegends.com/cdn/#{response['n']['summoner']}/data/en_GB/summoner.json" }
-      end
-
-      it 'calls the api' do
-        dd_summoner_spells
-        expect(ApiRequests).to have_received(:make_request).with(url)
-      end
-    end
-
-    describe '.dd_profile_icons' do
-      subject(:dd_profile_icons) { described_class.dd_profile_icons }
-      let(:url) do
-        { url: "http://ddragon.leagueoflegends.com/cdn/#{response['n']['profileicon']}/data/en_GB/profileicon.json" }
-      end
-
-      it 'calls the api' do
-        dd_profile_icons
-        expect(ApiRequests).to have_received(:make_request).with(url)
-      end
-    end
-
-    describe '.dd_specific_champion' do
-      subject(:dd_specific_champion) { described_class.dd_specific_champion(champion_name: champion_name) }
-      let(:champion_name) { 'Aatrox' }
-      let(:url) do
-        { url: "http://ddragon.leagueoflegends.com/cdn/#{response['n']['profileicon']}"\
-             "/data/en_GB/champion/#{champion_name}.json" }
-      end
-
-      it 'calls the api' do
-        dd_specific_champion
-        expect(ApiRequests).to have_received(:make_request).with(url)
-      end
-    end
-  end
-
   context 'region is invalid' do
     before do
-      allow(Starcall::Regions).to receive(:valid?).and_raise(Starcall::Regions::InvalidRegion)
+      allow(Starcall::Regions).to receive(:valid?).and_raise(Starcall::Error::InvalidRegion.new(region: region))
     end
 
-    describe '.dd_region_versions' do
-      subject(:dd_region_versions) { described_class.dd_region_versions }
-
-      it 'calls the api' do
-        expect { dd_region_versions }.to raise_error(StandardError)
-      end
-    end
-
-    describe '.dd_champions' do
-      subject(:dd_champions) { described_class.dd_champions }
-
-      it 'calls the api' do
-        expect { dd_champions }.to raise_error(StandardError)
-      end
-    end
-
-    describe '.dd_items' do
-      subject(:dd_items) { described_class.dd_items }
-
-      it 'calls the api' do
-        expect { dd_items }.to raise_error(StandardError)
-      end
-    end
-
-    describe '.dd_summoner_spells' do
-      subject(:dd_summoner_spells) { described_class.dd_summoner_spells }
-
-      it 'calls the api' do
-        expect { dd_summoner_spells }.to raise_error(StandardError)
-      end
-    end
-
-    describe '.dd_profile_icons' do
-      subject(:dd_profile_icons) { described_class.dd_profile_icons }
-
-      it 'calls the api' do
-        expect { dd_profile_icons }.to raise_error(StandardError)
-      end
-    end
-
-    describe '.dd_specific_champion' do
-      subject(:dd_specific_champion) { described_class.dd_specific_champion(champion_name: champion_name) }
-      let(:champion_name) { 'Aatrox' }
-
-      it 'calls the api' do
-        expect { dd_specific_champion }.to raise_error(StandardError)
+    describe '#initialize' do
+      let(:region) { 'not_a_region' }
+      it 'raises an error in initialisation' do
+        expect { described_class.new(region: region) }.to raise_error(Starcall::Error::InvalidRegion)
       end
     end
   end